got arcadedrive working and imported project to 2021. Please update your wpilib's!

+++ src/main/java/frc/robot/subsystems/Climber.java
@@ -96,7 +96,7 @@ public class Climber extends SubsystemBase {
   }
 
   public int getEncoderPos() {
-    return climberLeader.getSelectedSensorPosition();
+    return (int) climberLeader.getSelectedSensorPosition();
   }
 
   public double getEncoderVel() {
